requirements.txt
neonize>=0.3.14
python-dotenv
